requirements.txt
gradio
google-cloud-storage
google-cloud-speech
google-cloud-texttospeech
langchain-google-genai
langgraph
pydub
ffmpeg-python
